@@ -1,57 +1,63 @@
 {
 	"name": "pastebin",
 	"version": "0.0.1",
-	"description": "A pastebin clone made with SvelteKit and MongoDB",
-	"type": "module",
 	"private": true,
 	"scripts": {
-		"start": "vite build && node ./server/server.js",
-		"server": "node ./server/server.js",
+		"start": "vite build && vite start --host",
 		"dev": "vite dev --host",
-		"build": "vite build"
+		"build": "vite build",
+		"server": "bun ./server/server.js"
+	},
+	"type": "module",
+	"devDependencies": {
+		"@skeletonlabs/skeleton": "^3.0.0-next.10",
+		"@skeletonlabs/skeleton-svelte": "^1.0.0-next.19",
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/kit": "^2.17.2",
+		"@sveltejs/vite-plugin-svelte": "^4.0.4",
+		"@tailwindcss/container-queries": "^0.1.1",
+		"@tailwindcss/forms": "^0.5.10",
+		"@tailwindcss/typography": "^0.5.16",
+		"autoprefixer": "^10.4.20",
+		"prettier": "^3.5.1",
+		"prettier-plugin-svelte": "^3.3.3",
+		"svelte": "^5.20.2",
+		"svelte-check": "^4.1.4",
+		"tailwindcss": "^3.4.17",
+		"typescript": "^5.7.3",
+		"vite": "^5.4.14"
 	},
 	"dependencies": {
-		"@codemirror/commands": "^6.3.3",
+		"@auth/mongodb-adapter": "^3.7.4",
+		"@auth/sveltekit": "^1.7.4",
+		"@codemirror/commands": "^6.8.0",
 		"@codemirror/lang-angular": "^0.1.3",
 		"@codemirror/lang-cpp": "^6.0.2",
-		"@codemirror/lang-css": "^6.2.1",
-		"@codemirror/lang-html": "^6.4.7",
+		"@codemirror/lang-css": "^6.3.1",
+		"@codemirror/lang-html": "^6.4.9",
 		"@codemirror/lang-java": "^6.0.1",
-		"@codemirror/lang-javascript": "^6.2.1",
+		"@codemirror/lang-javascript": "^6.2.3",
 		"@codemirror/lang-json": "^6.0.1",
-		"@codemirror/lang-markdown": "^6.2.3",
+		"@codemirror/lang-markdown": "^6.3.2",
 		"@codemirror/lang-php": "^6.0.1",
-		"@codemirror/lang-python": "^6.1.3",
+		"@codemirror/lang-python": "^6.1.7",
 		"@codemirror/lang-rust": "^6.0.1",
 		"@codemirror/lang-sass": "^6.0.2",
-		"@codemirror/lang-sql": "^6.5.5",
+		"@codemirror/lang-sql": "^6.8.0",
 		"@codemirror/lang-vue": "^0.1.3",
-		"@codemirror/lang-xml": "^6.0.2",
-		"@codemirror/language": "^6.10.0",
-		"@codemirror/state": "^6.4.0",
+		"@codemirror/lang-xml": "^6.1.0",
+		"@codemirror/language": "^6.10.8",
+		"@codemirror/state": "^6.5.2",
 		"@codemirror/theme-one-dark": "^6.1.2",
-		"@codemirror/view": "^6.23.0",
-		"@lezer/highlight": "^1.2.0",
+		"@codemirror/view": "^6.36.3",
+		"@iconify/svelte": "^4.2.0",
+		"@lezer/highlight": "^1.2.1",
 		"@replit/codemirror-lang-svelte": "^6.0.0",
-		"bcrypt": "^5.1.1",
-		"body-parser": "^1.20.2",
-		"cm6-graphql": "^0.0.9",
+		"@sveltejs/adapter-vercel": "^5.6.3",
+		"cm6-graphql": "^0.2.0",
 		"codemirror": "^6.0.1",
-		"cors": "^2.8.5",
-		"dotenv": "^16.3.1",
-		"express": "^4.18.2",
-		"express-fileupload": "^1.4.3",
-		"mongoose": "^7.6.8"
-	},
-	"devDependencies": {
-		"@iconify/svelte": "^3.1.6",
-		"@sveltejs/adapter-node": "^1.3.1",
-		"@sveltejs/kit": "^1.30.3",
-		"@types/bcrypt": "^5.0.2",
-		"svelte": "^4.2.8",
-		"svelte-check": "^3.6.3",
-		"tslib": "^2.6.2",
-		"typescript": "^5.3.3",
-		"vite": "^4.5.1"
+		"mongodb": "^6.13.1",
+		"mongoose": "^8.10.1",
+		"pastebin": "file:"
 	}
 }
