@@ -1,37 +1,12 @@
-// @ts-nocheck
-import fs from "fs";
+import type { PageServerLoad } from './$types';
 
-import images from "$lib/ts/mongo/image";
-import users from '$lib/ts/mongo/user';
+import { db } from '$lib/ts/db';
 
-/** @type {import('./$types').Actions} */
-export const actions = {
-	image: async ({ cookies, request }) => {
-		let user = null;
-		if (cookies.get('authKey', { path: '/' })) {
-			user = await users.get(cookies.get('authKey', { path: '/' }));
-			user.password = null;
-		}
+export const load: PageServerLoad = async ({ locals }) => {
+    const session = await locals.auth();
 
-		let formData = await request.formData();
+    if(!session) return { session };
+    let bins = await db.getBins({ email: session?.user?.email });
 
-		const data = Object.fromEntries(formData);
-
-		try {
-			await fs.writeFileSync(`./images/${user._id}/${formData.get("file").name}`, Buffer.from(await (data.file as Blob).arrayBuffer()), "base64");
-		} catch (e) {
-			
-			await fs.mkdirSync(`./images/${user._id}`, { recursive: true });
-	
-			try {
-				await fs.writeFileSync(`./images/${user._id}/${formData.get("file").name}`, Buffer.from(await (data.file as Blob).arrayBuffer()), "base64");
-			} catch (e) {
-				return;
-			}
-		}
-
-		await users.addImage(user._id, formData.get("file").name);
-    
-		await images.create(formData.get("file").name, formData.get("file").type, formData.get("file").size);
-	}
+    return { session, bins };
 };
