@@ -2,8 +2,5 @@ import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
-	plugins: [sveltekit()],
-	server: {
-		port: 5000
-	},
+	plugins: [sveltekit()]
 });
